Pass in correct style of pedantic when parsing tables
The old way was pedantic=True, the new way is verify='exception',
although it has more options.

Also make all the callers use the astropy version of raise or warn,
because then it will be consistent with the calling parameters.
Right now, we're calling both astropy and a local version.

Fixes #130.