Fix regtap doc warnings
Properties section isn't rendered right now and it emits a warning,
because docutils doesn't know what to do with it.  Looking at the
rendered 0.9.2 page it looks like it wasn't being displayed anyway,
but all the properties after it are rendered with their own doc
strings.  So let's get rid of that second.

The warnings about emphasis were a misparsing of python arg type
stuff, so I have specifically marked that as python code.