Result for TAP service should default to id=result
For a TAP service async job, it can have multiple results in one
set.  When doing a fetch_results() it has to pick which one of the
multiple results to fetch.  The TAP specification outlines that
if there is a query language that returns on result (as ADQL does),
then the result of the query should have id of result in the XML
representation of the results.

Right now pyvo picks just the first one in the list, which doesn't
play well with the CADC TAP service, which has the real result URL
listed last.  If no result with id=result is present, then we
default to our previous behavior of picking the first one.  In
these cases it is likely that the user will need to fetch all
the results through a different mechanism.

Fixes #147.