Refactor session to not be a singleton
Right now, session was a singleton, and there was only one of
them and all the calls were using it.  This is great from
simplicity point of view, but for auth, you need more context
to set the correct auth methods on each call.  This context
will end up being a part of the session, which also makes the
call.

This commit creates the session at the service level, and passes
it down into all the subclasses it creates, such as queries,
results, and records.