Fix AstropyDeprecationWarning in SIA & SSA
Fixes #163

When running SSAService.search() and passing in an astropy Quantity,
this warning would be emitted:

WARNING: AstropyDeprecationWarning: The truth value of a Quantity is ambiguous.
In the future this will raise a ValueError. [astropy.units.quantity]

Instead of checking the truth value of the parameter, I'm just going to check
if it is None or not, not relying on what the truth value of the Quantity
may be.