Fix unit test that doesn't pass on Windows
When enabling Windows unit testing, we found a bug in the path
logic where it was assuming that it was going to deal with paths
in the posix way ('/'), but just used os.path.  This breaks for
windows, where os.path.join will use '\' instead.  So just force
this to doing posix style, not local implied style.