Fix docstring build error
From Travis:
Warning, treated as error:
/home/travis/build/astropy/pyvo/build/lib/pyvo/io/uws/tree.py:docstring of pyvo.io.uws.tree.Parameters:22:Unexpected section title or transition.

Looks like sphinx is confusing Parameters (the name of the class)
for a heading of where the parameters to a function call should be.

This happens even if I get rid of all the docstrings.  So instead,
since this is a fairly internal class, I'm going to rename the class
to ParametersElement.