Apply patch from astropy/astropy@05a62450ce2ec7a75328837a47f9452ace59b2c9
[PATCH] BUG: Fix missing server_close method on SAMPHubProxy

It's possible this bug has existed for a long time and was just hidden.
Another possibility is it appeared at some point during refactoring, but
as far as I can tell it's been around a long time.

The  attribute is still just a ; it
does not make sense for it to have a  method because it
does not itself run or wrap a server.

It does however manage a pool of XML-RPC client connections which is
deleted when calling .  To be on the safe
side I added a  method to the proxy pool
class which closes the client connections.  This is called when shutting
down the  server loop.  Unlike the normal
class, the  used only in tests never runs a TCP server
of its own, so calling  anywhere in here does not make
sense.

Co-authored-by: E. Madison Bray <erik.bray@lri.fr>