Column selection and grouping for datasearch.
This introduces expected_columns in existing regtap.RegistryResource,
which I now intend to re-use (and evolve) for the data discovery, too.
It will now declare what columns are required from the queries itself
(where I suspect we'll have to be a bit niftier if we want more
advanced categories of data we get in the initial discovery query).