Fixing RegistryResource.describe for multi-capability resources.
The previous code assumed that typically, records only have one capability,
which is no longer true as soon as people do not constrain by servicetype
(actually, it hasn't been true before either).

So, describe() now gives the various access methods and no longer tries
to indicate a specific capability.  I have re-shuffled the order of
the outputs a bit in consequence.

And I wonder if this should show contact information by default.  The
main counterargument at this point: That requires a network request.