Canonicalize case of TAP query parameter id
When retrieving the contents of a TAP query from an AsyncTAPJob,
lowercase the identifier of the parameter before comparing it to
"query". Some TAP servers reflect the case of the parameter in the
job XML and DALI says that parameter keys are case-insensitive, so
the parameter ID may be "QUERY" or some other case.

test_submit_job was using the wrong arguments to submit_job()
(passing a URL first instead of the query), uncovered by testing
this change.  Fix that as well.

In MockAsyncTAPServer, stop lowercasing incoming parameters and
instead reflect the provided case in the parameter ID, making it
possible to test this fix.

Fixes #356