Unrelated: doctest fix.
The tables in resource.get_tables() come in random column order (because
they are fed from the results of relational database queries).  That's
arguably not ideal, but it is what we can do given what the registry
has in terms of metadata, and it actually doesn't hurt as long as people are
aware of it.

This commit changes an example in the documentation to become reproducible.
The example gets a bit uglier this way, but it's probably still on the
good side of making examples runnable rather than didactic (where the
two conflict, which they do in this case).