The Freetext condition now adapts to target server.
More specifically, it will use UNION only if available, falling back to
the much-harder-to-plan OR otherwise.

This commit also fixes a bug in the UNION-based condition where multiple
words were joined (in effect) with an OR rather than and AND as
documented.

To make this more convenient, this adds a get_tap_cap method to
TAPServices; this returns the (first) tr:TableAccess capability on the
service.  The IVOA are not really saying what multiple such capabilities
would mean, so for now I'd say we're good on just returning the first
one (and it's in line with what we've been doing so far with maxrec and
hardlimit; upload_methods would now change behaviour if there were
upload methods distributed over multiple TableAccess-es; but the
previous behaviour would have been bad anyway).

We are also adding convenience methods for accessing the tap capability
to the TAPService, for accessing the ADQL language in there, and for
pulling out language features and UDFs.  We probably want to document
these somewhere visible, as advanced users (cross-server) can do smart
things with them.  But then perhaps that's too esoteric for now.

It is an interesting question whether at least get_adql (or something
more aptly named) should sit on TAPService itself...