Making the all-VO query example work to some extent.
Querying for dataproduct_type was a bad idea as we get back all kinds
of strings (and sometimes NULLs, too), which vstack() refuses to
merge.

We're now using a query with floats, which should be a lot less fragile.
At least for the first few services, this now runs.  It will probably
fail when a service returns NULLs for s_ra and s_dec, though.