Making SIA2 discoverable as intended by the standard.
This introduces a new sort of syntax for the Servicetype constraint, as
the "new-style" standard ids must be matched ignoring the minor version.
*If* further standards follow SIA 2's pattern, we'll have to think about
refactoring this.  For now, I'll argue for returning to the previous
pattern.

With SIA2, there's the additional problem that the authors re-used the SIA1
main identifier.  That's a mortal sin as far as our standard id ->
service class resolution is concerned.  We hence fix this on digesting
the query results coming from the registry.  I'm pretty sure nothing like
this will happen again.  Let's consider it a fluke on SIA2.

This commit as an integration test-type test at the end of test_regtap.
Perhaps we should move this to test_sia2_remote and replace one of the
rather slow tests there?  I'd like to leave that call to someone else.

(also, no longer checking E128 and E131, i.e., visual indents, in tox)