UWS messages now sit on errorsummary.
Previously, they were parsed into job.message, which was probably unintended,
as there were message attributes on errorsummary that just were never
filled because errorsummary would not parse its children itself.

The alternative would be to preserve the previous behaviour, where message
sits on the job element; but that's against the UWS schema, and if we
did that, we'd have to take away the extra attributes on errorsummary
(and revert the change to raise_if_error in the TAPJob).

If we do it like this, that would fix Bug #431.