Deprecating ivoid2service.
The problem with it is that when a resource has multiple capabilities
(e.g., SCS and TAP, as is now true for the big majority of VO resources),
its result is ill-defined.

One *could* add a parameter to select the kind of service desired, but
that's complicated in detail.  Hence, we advise people to use normal
registry search with an ivoid constraint.

This is supposed to address Bug #425.