Deprecating image and spectrum as registry servicetype-s
In an attempt to accomodate the concerns brought up in
https://github.com/astropy/pyvo/pull/449#issuecomment-2230934353,
we also explain a bit why image and spectrum are being deprecated.
we still raise deprecation warnings (rather than something else), as
that is, I think, what we are doing (and should be doing): We are saying
that we discourage people to use these two strings and reserve the
right to drop them one day.