Making a first global discovery query work.
This changes the internal parameter processing in _ImageDiscoverer to
plain floats instead of quantities (which turned out to be painful with SIA1)

We also change the naming scheme for LearnableRequestMockers because
the old one was not producing unique names.

Also, joined tables in regtap are now alphabetically sorted to get more
reproducable RegTAP queries, which we

Finally, we have to do a quick fix to SIA2, which currently is broken with the
registry in main; this should be backed out when sia2 is fixed in main.