global discovery: get_service is now lax.
This is relevant when, for one reason or other, multiple capabilities
of a type are present in a registry record.  For our purposes here
(sia, ssap, tap...), it is highly unlikely that these different capabilities
correspond to different data (that's as different from VizieR SCS, where
different capabilities correspond to different tables; but that's a pain
and we don't want anything like that here).