Updating global discovery test cases.
This is mainly about using 1=contains rather than distance in the obscore
queries; we also had to update the RegTAP capability.xml in the registry
test to bring it in line with what the discover tests do.  A better
isolation of these test suites would be desirable; but then we don't want
forsake capabilities caching altogether either...

Also, sneaking in a few formatting problems.