Fixing SODA_SYNC_IVOID so we work with all version 1 SODA services.
While protocol IVOIDs in the VO give the full version, matching must
ignore the the minor version or we'll break when a new minor version is
out.  Discussion on version 1.1 of SODA is just starting.  Let's better
clean this up now.

Actually, it would even be better if we fixed get_adhocservice_by_ivoid
to accept a regular expression, because we we really ought to be matching
aginst is ivo://ivoa.net/std/SODA#sync-1\.\d+; but I'd say that can
wait until there is thought of SODA 10.x.