Avoid assuming that `access_url` exists always
DatalinkResultsMixin was assuming that the records created by any
subclass would have the `access_url` attribute. This is not true for SSA
(whether it should do is a separate question). This replaces that
assumption with a check for existence first, and then if that fails,
fall back to getdatalink on the record.