SodaRecordMixin no longer will accept a datalink#links endpoint for soda.
It is unclear why it ever did this, as I really do not see how this would
ever have a beneficial effect (the only parameter understood by both links and
SODA is ID, and even that is interpreted in totally different ways).  It
has, however, a detrimental effect; see bug #578.