Now accepting sia1 as an alias for sia in registry.Servicetype.
That's so there is a symmetry to sia2.  When further version 2-standards
appear, I'd suggest we continue that pattern: The official IVOA short name
(without numbers in version 1) remains, but a preferred alias with a 1 is
added in pyVO.

Note that access_modes and friends do not follow this logic; for SIA1, they
still do the regular thing and use the IVOA short name.

Would it make sense to convince the IVOA to add the 1 to their short names?
Well, it would have all these years ago.  Now... I think it would just
increase the confusion.