Dropping the UAT vocabulary fixture.
It was a bad idea to have a compact UAT as a local resource; it poisons
astropy's download cache and is then hard to get rid of again, in particular
because pyvo caches the vocabularies in RAM.

In consequence, the unit tests for the UAT constraint are now marked as
using the network.  It does, too, when you run the tests in a freshly made
environment.

I'd still like to keep the commit with the fixture in the history.  Perhaps
we want something like that elsewhere some day.